backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.4.0
redis==5.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
cryptography==42.0.5
scikit-learn==1.4.1
tensorflow==2.15.0
web3==6.15.1
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
Pillow==10.2.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.2
face-recognition==1.3.0
pydantic==2.5.2 